Extract `Saver` from `Rotator`.
