+++ src/app-util/index.js
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from '#x/Rotator';
+export * from '#x/Saver';
